my-finance/apps/ml-service/python-ml/requirements.txt
# ML Dependencies
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0

# gRPC Server
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# API Server (optional, for REST fallback)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Utils
python-dotenv>=1.0.0
